Create ssx-core package, populate and update other packages
