+++ packages/ssx-core/.eslintignore
@@ -1,0 +1,3 @@
+dist
+node_modules
+header.png
