+++ packages/ssx-core/.npmignore
@@ -1,0 +1,6 @@
+example
+.editorconfig
+.eslintignore
+.eslintrc
+.prettierrc
+header.png
