+++ packages/ssx-core/README.md
@@ -1,0 +1,13 @@
+# SSX-Core
+
+SSX-Core is a library made to aggregate types and utils to other SSX packages.
+
+## Installation
+
+You can add SSX-Core from yarn or npm:
+
+```bash
+npm install @spruceid/ssx-core
+# or
+yarn add @spruceid/ssx-core
+```
