+++ packages/ssx-core/package.json
@@ -1,23 +1,20 @@
 {
-  "name": "@spruceid/ssx-serverless",
+  "name": "@spruceid/ssx-core",
   "version": "1.0.0",
-  "description": "Create SSX applications in serverless architectures",
+  "description": "SSX core library",
   "author": "Spruce Systems Inc.",
   "license": "Apache-2.0 OR MIT",
-  "homepage": "https://github.com/spruceid/ssx-serverless#readme",
+  "homepage": "https://github.com/spruceid/ssx/packages/ssx-core#readme",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "bugs": {
-    "url": "https://github.com/spruceid/ssx-serverless/issues"
+    "url": "https://github.com/spruceid/ssx/issues"
   },
   "directories": {
     "src": "src"
   },
   "scripts": {
-    "build": "tsc",
-    "doc": "yarn doc:extractor && yarn doc:documenter",
-    "doc:extractor": "api-extractor run",
-    "doc:documenter": "api-documenter generate -i temp -o ../../documentation/reference/ssx-serverless"
+    "build": "tsc"
   },
   "engines": {
     "node": ">=16.13.0"
@@ -27,14 +24,12 @@
     "url": "git+https://github.com/spruceid/ssx.git"
   },
   "dependencies": {
-    "siwe": "^2.1.2",
-    "@spruceid/ssx-core": "*",
-    "@spruceid/ssx-gnosis-extension": "*",
-    "axios": "^0.27.2"
+    "axios": "^0.27.2",
+    "express": "^4.18.1",
+    "express-session": "^1.17.3",
+    "siwe": "^2.1.2"
   },
   "devDependencies": {
-    "@microsoft/api-documenter": "^7.19.4",
-    "@microsoft/api-extractor": "^7.29.2",
     "@typescript-eslint/eslint-plugin": "^5.23.0",
     "@typescript-eslint/parser": "^5.23.0",
     "eslint": "^8.15.0",
