+++ packages/ssx-gnosis-extension/package.json
@@ -15,6 +15,9 @@
     "build": "tsc",
     "clean": "rimraf dist"
   },
+  "dependencies": {
+    "@spruceid/ssx-core": "*"
+  },
   "peerDependencies": {
     "siwe": "^2.1.2",
     "ethers": "^5.7.1"
