+++ packages/ssx-sdk/src/ssx.ts
@@ -4,12 +4,12 @@ import {
   SSXInit,
 } from './core';
 import {
-  SSXConfig,
-  SSXSession,
+  SSXClientConfig,
+  SSXClientSession,
   SSXRPCProviders,
   SSXEnsData,
   SSXEnsResolveOptions,
-} from './types';
+} from '@spruceid/ssx-core';
 
 declare global {
   interface Window {
@@ -17,7 +17,7 @@ declare global {
   }
 }
 
-const SSX_DEFAULT_CONFIG: SSXConfig = {
+const SSX_DEFAULT_CONFIG: SSXClientConfig = {
   providers: {
     web3: {
       driver: globalThis.ethereum,
@@ -30,11 +30,11 @@ const SSX_DEFAULT_CONFIG: SSXConfig = {
  * A toolbox for user-controlled identity, credentials, storage and more.
  */
 export class SSX {
-  /** SSXSession builder. */
+  /** SSXClientSession builder. */
   private init: SSXInit;
 
   /** The session representation (once signed in). */
-  public session?: SSXSession;
+  public session?: SSXClientSession;
 
   /** Current connection of SSX */
   public connection?: SSXConnected;
@@ -42,7 +42,7 @@ export class SSX {
   /** Supported RPC Providers */
   public static RPCProviders = SSXRPCProviders;
 
-  constructor(private config: SSXConfig = SSX_DEFAULT_CONFIG) {
+  constructor(private config: SSXClientConfig = SSX_DEFAULT_CONFIG) {
     this.init = new SSXInit({ ...this.config, providers: { ...SSX_DEFAULT_CONFIG.providers, ...this.config?.providers } });
 
     if (this.config.enableDaoLogin) {
@@ -55,7 +55,7 @@ export class SSX {
    * Request the user to sign in, and start the session. 
    * @returns Object containing information about the session
    */
-  async signIn(): Promise<SSXSession> {
+  async signIn(): Promise<SSXClientSession> {
     try {
       this.connection = await this.init.connect();
     } catch (err) {
