+++ packages/ssx-server/package.json
@@ -23,6 +23,7 @@
     "ssx-server": "bin/ssx-server.js"
   },
   "dependencies": {
+    "@spruceid/ssx-core": "*",
     "@spruceid/ssx-gnosis-extension": "*",
     "axios": "^0.27.2",
     "body-parser": "^1.20.0",
